requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
pypdf==4.3.1
markdown==3.6
openai==1.43.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
